Rename integration test files from main to something specific
