+++ crates/fs/Cargo.toml
@@ -15,7 +15,7 @@ test = false
 [[test]]
 name = "integration"
 required-features = ["test-support"]
-path = "tests/integration/main.rs"
+path = "tests/integration/fs_tests.rs"
 
 [dependencies]
 anyhow.workspace = true
