--- crates/fs/tests/integration/main.rs
@@ -1,2 +1,0 @@
-mod fake_git_repo;
-mod fs;
