+++ crates/worktree/Cargo.toml
@@ -13,7 +13,7 @@ test = false
 [[test]]
 name = "integration"
 required-features = ["test-support"]
-path = "tests/integration/main.rs"
+path = "tests/integration/worktree_tests.rs"
 
 [lints]
 workspace = true
