+++ crates/worktree/tests/integration/worktree_settings_tests.rs
